requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
Django==6.0
django-environ==0.12.0
djangorestframework==3.16.1
factory_boy==3.3.3
Faker==38.2.0
mysqlclient==2.2.7
pillow==12.0.0
pycparser==2.23
pillow==12.0.0
sqlparse==0.5.4
tzdata==2025.2
